frontend starter

+++ backend/controllers/product.controller.js
@@ -62,6 +62,13 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  // catching the 404 error
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ sucess: false, messsage: "Product ID Not Found" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -69,6 +76,6 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log(`Error in deleting productr: ${error.message}`);
-    res.status(404).json({ success: false, message: "Product Not Found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
